Implement configure.py to generate ninja build script
While nothing is containerized in Docker yet, this commit implements the
workflow that will run inside the container to build projects. You can
run it manually to check it out if you want, but it's a big pain.

You'll need:
    - An input project with a Hi-C file and workflow.yaml
      (calling it project.yaml instead is also acceptable)
    - The hic2structure.py script installed and available
      in your PATH
    - The Ninja build tool installed on your system
    - LAMMPS installed on your system (called 'lmp')
    - The 4DGB Browser repository cloned and checked-out
      somewhere on your system (along with gtk.min.js built
      and placed in the repo's server/static/gtk/js/ directory)

(This is all stuff that would normally be handled when the Docker
image is built)

From there, you can run the scripts/configure.py script with the
following arguments:

    scripts/configure.py path/to/project/directory path/to/output/directory path/to/4DGB/repo

After running that, switch the the output directory and run 'ninja' to build the
project. This will take a while since it needs to run the LAMMPS simulation.
Once done, just run 'python3 gtkserver.py' to launch a browser instance.