Implement containerized workflow
This is, for the most part, a copy of the functionality that's already
been implemented in the 4DGB Browser repository (with the plan being
to remove most of that functionality from the other repo).

The included Dockerfile gathers all the dependencies needed to build
a project from the input data, including cloning and building other
required repos.

The entrypoint script for the Docker image does most of the work
of actually building the project (using Ninja for incremental
builds) and then starting an instance of the browser.

With an image built, users only need to run the 'run_project' script
and provide it with the path to their project directory. This script
calls 'docker run' with all the appropriate arguments so that their
project will be built and made available to their web browser.