Setup dependencies as git submodules, implement project parameters
This commit adds the dependency repos as git submodules, which are
then copied into the Docker build as opposed to performing the clone
within the Dockerfile. The hope is that this will make iteration
quicker since we can now make changes to the other repos and build
images with those changes without having to commit/push first.

I also added support for the 'interval', 'threshold' and 'chromosome'
parameters. These are specified in the workflow.yaml and are used
as arguments to the hic2structure.py script.