Refactor workflow, add parameters
Using Ninja to handle building the project was a fun expiriment, but it ended up
just making things more complicated in the end (and we couldn't get useful log
output from it). This commit refactors the workflow so that it's all handled
by a big script in scripts/workflow.py. This new script is big, but is infinitely
more readable and flexible than the previous configure.py script(s).

However, this refactor does not yet handle track data. We'll be integrating
a script to manage that soon, but for now, this commit will be on its own
branch until it's ready to totally replace the main branch.

This commit  also incorporates two new settings for configuring LAMMPS runs:
'timesteps' and 'bond_coeff', which can be specified in the workflow.yaml
file.