Integrate csv2tracks script into workflow
The workflow script now calls the new csv2tracks script to generate track
data from csv files as a part of the main workflow. The example project
has been updated to use this.

This required a few small changes to the csv2tracks script:
    - Columns are referenced by index rather than name in several
      places to fix cases where columns in different files had the same
      name.
    - Added a "--relative" argument to make the paths in the output
      metadata json files relative to a different directory.

To celebrate, the version number has been incremented.