New & Improved runner script
This commit introduces a newly re-written runner script, now called
'4DGBWorkflow'. It's in python3 now, rather than bash, but it sticks
only to what's in the (thankfully, massive) standard library, so
it should still be pretty portable. I tested it with all major
python versions from 3.7 to 3.10 and it worked fine.

This new command has several subcommands, along with many
optional flags to tweak how it works: 'run' does what the
previous script did: run the docker container. 'update' does
a 'docker pull' to fetch a new version of the image, 'version'
reports the version of the workflow and the browser, and 'template'
generates a copy of the example project.

About that example project, the script 'generate_blob.py' is used
to create a hunk of base85 text from a tarball of the example_project.
This is then copy-pasted into the 4DGBWorkflow script and used to
create the template directory.